remote-sensing-agent/requirements.txt
google-cloud-storage
google-cloud-bigquery
numpy
rasterio
rioxarray
xarray
shapely
dask